requirements.txt
Django==2.0
djangorestframework==3.8.2
djangorestframework-jsonapi==2.5.0
djangorestframework-datatables